feat(query): Phase 1.4 OQL end-to-end — decoder / Struct / extensions / tests
ChunkedQueryResponse<T> 完整解碼（cppcache RemoteQuery.cpp:3291-3480 鏡像）:
- HandleChunk C1-C12: chunk-part-header 分流（Exception / NullObject /
  Object）, skipClass parent metadata, 讀 collection type name, 處理
  Struct fieldNames（含 cross-chunk dedup）, 兩條 array-type 分支
  CacheableObjectArray(52) / CacheableObjectPartList(25)
- ReadObjectPartList R1-R3 + ReadStructRow / ReadExceptionAndThrow
  helpers — 內層 Struct row 用 local buffer 收 K 個 field values，
  直接在 collector 組成 QueryStruct push（Option C），跳過 cppcache
  的 "flat -> outer reshape" 中介
- SkipClass S1-S4: 讀 DSCode.Class + typeId + u16 classLen + advance
- ReadShortString helper: 同時接 CacheableString(42) / CacheableASCIIString(87)
  兩種 DSCode 短字串形式（server 對純 ASCII 名稱實際送 87）

QueryStruct 公開型別（拉前自 Phase 2）:
- IReadOnlyList<object?> + by-name indexer + FieldNames / GetFieldIndex
  / GetFieldName。命名避開 C# struct keyword
- StructSet 路徑：collector 內每 K 個 row values 組好 QueryStruct push，
  B10 簡化成單行 return

QueryExtensions:
- ExecuteSingleAsync / ExecuteFirstOrDefaultAsync — scalar 包裝
- WithParameters / WithResponseTimeout — fluent setter（鏈式 IQuery<T>）

NewQuery<T> type guard:
- T 必須是 SerializationRegistry 註冊型 或 QueryStruct
- 擋掉 bucket 2 (PDX 自訂類，Phase 2) / bucket 4 (ORM mapping)
- SerializationRegistry.IsRegistered(Type) 新 API

BigEndianBinaryReader.ReadArrayLength:
- Java 變長 array length 解碼（cppcache DataInput::readArrayLength 對等）
- 1 byte / 0xFE+u16 / 0xFD+i32 / 0xFF=-1

兩個整合測試實戰抓到的真實 bug:

1. TcrMessageHelper.ReadChunkPartHeader 簽號 byte 錯解 — DSFid 負值
   (CollectionTypeImpl = -59, 0xC5) 用 unsigned ReadByte 讀回 197 跟 -59
   比對失敗。修法 `compId = (sbyte)reader.ReadByte()`。之前 GetAll /
   RemoveAll 等用正 DSFid 沒爆過；query 是第一個碰到負 DSFid 的路徑。

2. ChunkedQueryResponse C6/C7/R3a 對短字串 DSCode 太嚴 — 原本只接
   CacheableString(42)，server 對純 ASCII 類別名 / 欄位名實際送
   CacheableASCIIString(87)。抽 ReadShortString 同時接兩種。

OQL 寫法:
- `this` 在 WHERE clause 不 work（至少對 int region），整合測試一律
  用顯式 alias `SELECT t FROM /test t WHERE t = ...`

測試:
- 單元 +39: QueryStructTests(16) / QueryExtensionsTests(18) /
  TcrMessageBuilderQueryTests(17) / TcrMessageBuilderQueryWithParametersTests(22)
- 整合 +7: QueryIntegrationTests 全部 PASS（SELECT * / COUNT / params /
  ExecuteSingleAsync / type mismatch）

Phase 1.4 剩餘:
- Region.ExistsValueAsync / SelectValueAsync
- 多欄 projection 整合測試（需要 server-side PDX 結構化資料）

+++ src/Geode.Client/Protocol/TcrMessageHelper.cs
@@ -155,7 +155,13 @@ internal sealed class TcrMessageHelper(ILogger<TcrMessageHelper> logger)
             }
             else if (expectedDsCode == DSCode.FixedIDByte)
             {
-                compId = reader.ReadByte();
+                // DSFid is a signed byte on the wire; cppcache reads it
+                // via int8_t. Without the (sbyte) cast 0xC5 reads back
+                // as 197 (unsigned) instead of -59 (CollectionTypeImpl),
+                // breaking the compId compare. Only matters for negative
+                // DSFid IDs — the positive ones (VersionedObjectPartList,
+                // CacheableObjectPartList, etc.) round-trip either way.
+                compId = (sbyte)reader.ReadByte();
             }
         }
